# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CS546 API",
        "description": "Connecting clothing to a smart forum",
        "version": "0.0.1"
    },
    "host": "localhost:5000",
    "schemes": [
        "http"
    ],
    "basePath": "/",
    "produces": [
        "application/json",
        "text/html"
    ],
    "paths": {
...
        "/profile/": {
            "get": {
                "summary": "User Profile",
                "description": "The User Profile endpoint returns information about the Uber user that has authorized with the application.",
                "tags": [
                    "User"
                ],
                "responses": {
                    "200": {
                        "description": "Profile information for a user",
                        "schema": {
                            "$ref": "#/definitions/profile"
                        }
                    },
                    "default": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/error"
                        }
                    }
                }
            },
            "put": {
                "summary": "Update User Profile",
                "desciption": "Allow the user to update fields in their user profile",
                "tags": [
                    "User"
                ],
                "parameters": [ 
                    {
                        "name": "display_name",
                        "in": "header",
                        "desciption": "New display name for user",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "header",
                        "desciption": "New email for user",
                        "required": false,
                        "type": "string"
                       
                    },
                    {
                        "name": "password",
                        "in": "header",
                        "desciption": "New password for user",
                        "required": false,
                        "type": "string"
                        
                    },
                    {
                        "name": "isMale",
                        "in": "header",
                        "desciption": "New gender for user",
                        "required": false,
                        "type": "string"
                        
                    },
                    {
                        "name": "avatar",
                        "in": "header",
                        "desciption": "New picture for user",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "desciption": "Profile info updated"
                    },
                    "500": {
                        "desciption": "Internal server error"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "profile": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "string",
                    "description": "id of the user."
                },
                "display_name": {
                    "type": "string",
                    "description": "Display name of the user."
                },
                "email": {
                    "type": "string",
                    "description": "Email address of the user"
                },
                "isMale": {
                    "type": "boolean",
                    "desciption": "Gender of the user"
                },
                "interestedForums": {
                    "type": "array",
                    "item": {
                        "type": "string",
                        "desciption": "forum id"
                    }
                },
                "avatar": {
                    "type": "string",
                    "description": "Image URL of the user in base64."
                }
            }
        },
        "error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int32"
                },
                "message": {
                    "type": "string"
                },
                "fields": {
                    "type": "string"
                }
            }
        }
    }
}